+++ log/zerolog/doc.go
@@ -1,0 +1,2 @@
+// Package zerolog provides a log.Logger implementation backed by github.com/rs/zerolog.
+package zerolog